+++ pkg/query/logical/measure/measure_plan_indexscan_local.go
@@ -27,7 +27,7 @@ import (
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 	measurev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/measure/v1"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
-	"github.com/apache/skywalking-banyandb/pkg/index"
+	"github.com/apache/skywalking-banyandb/pkg/index/inverted"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
 	pbv1 "github.com/apache/skywalking-banyandb/pkg/pb/v1"
 	"github.com/apache/skywalking-banyandb/pkg/query"
@@ -87,7 +87,7 @@ func (uis *unresolvedIndexScan) Analyze(s logical.Schema) (logical.Plan, error) 
 		// fill AnyEntry by default
 		entity[idx] = pbv1.AnyTagValue
 	}
-	filter, entities, err := logical.BuildLocalFilter(uis.criteria, s, entityMap, entity, true)
+	query, entities, _, err := inverted.BuildLocalQuery(uis.criteria, s, entityMap, entity)
 	if err != nil {
 		return nil, err
 	}
@@ -100,7 +100,7 @@ func (uis *unresolvedIndexScan) Analyze(s logical.Schema) (logical.Plan, error) 
 		projectionTagsRefs:   projTagsRefs,
 		projectionFieldsRefs: projFieldRefs,
 		metadata:             uis.metadata,
-		filter:               filter,
+		query:                query,
 		entities:             entities,
 		groupByEntity:        uis.groupByEntity,
 		uis:                  uis,
@@ -114,7 +114,7 @@ var (
 )
 
 type localIndexScan struct {
-	filter               index.Filter
+	query                *inverted.Query
 	schema               logical.Schema
 	uis                  *unresolvedIndexScan
 	order                *logical.OrderBy
@@ -155,7 +155,7 @@ func (i *localIndexScan) Execute(ctx context.Context) (mit executor.MIterator, e
 		Name:            i.metadata.GetName(),
 		TimeRange:       &i.timeRange,
 		Entities:        i.entities,
-		Filter:          i.filter,
+		Query:           i.query,
 		OrderByType:     orderByType,
 		Order:           orderBy,
 		TagProjection:   i.projectionTags,
@@ -172,7 +172,7 @@ func (i *localIndexScan) Execute(ctx context.Context) (mit executor.MIterator, e
 func (i *localIndexScan) String() string {
 	return fmt.Sprintf("IndexScan: startTime=%d,endTime=%d,Metadata{group=%s,name=%s},conditions=%s; projection=%s; order=%s;",
 		i.timeRange.Start.Unix(), i.timeRange.End.Unix(), i.metadata.GetGroup(), i.metadata.GetName(),
-		i.filter, logical.FormatTagRefs(", ", i.projectionTagsRefs...), i.order)
+		i.query, logical.FormatTagRefs(", ", i.projectionTagsRefs...), i.order)
 }
 
 func (i *localIndexScan) Children() []logical.Plan {